+++ Sources/StreamChat/Models/Attachments/AttachmentTypes.swift
@@ -1,5 +1,5 @@
 //
-// Copyright © 2023 Stream.io Inc. All rights reserved.
+// Copyright © 2024 Stream.io Inc. All rights reserved.
 //
 
 import Foundation
@@ -119,17 +119,13 @@ public struct AttachmentType: RawRepresentable, Codable, Hashable, ExpressibleBy
 }
 
 public extension AttachmentType {
-    /// Backend specified types.
     static let image = Self(rawValue: "image")
     static let file = Self(rawValue: "file")
     static let giphy = Self(rawValue: "giphy")
     static let video = Self(rawValue: "video")
     static let audio = Self(rawValue: "audio")
     static let voiceRecording = Self(rawValue: "voiceRecording")
-
-    /// Application custom types.
     static let linkPreview = Self(rawValue: "linkPreview")
-    /// Is used when attachment with missing `type` comes.
     static let unknown = Self(rawValue: "unknown")
 }
 
@@ -218,6 +214,8 @@ public enum AttachmentFileType: String, Codable, Equatable, CaseIterable {
     case mov, avi, wmv, webm
     /// Image
     case jpeg, png, gif, bmp, webp
+    /// Unknown
+    case unknown
 
     private static let mimeTypes: [String: AttachmentFileType] = [
         "application/octet-stream": .generic,
@@ -302,6 +300,10 @@ public enum AttachmentFileType: String, Codable, Equatable, CaseIterable {
         default:
             return false
         }
+    }
+
+    public var isUnknown: Bool {
+        self == .unknown
     }
 }