+++ Sources/StreamChat/Query/UserListQuery.swift
@@ -1,5 +1,5 @@
 //
-// Copyright © 2023 Stream.io Inc. All rights reserved.
+// Copyright © 2024 Stream.io Inc. All rights reserved.
 //
 
 import Foundation
@@ -114,7 +114,7 @@ public struct UserListQuery: Encodable {
     }
 }
 
-extension UserListQuery {
+public extension UserListQuery {
     /// Builds `UserListQuery` for a user with the provided `userId`
     /// - Parameter userId: The user identifier
     /// - Returns: `UserListQuery` for a specific user