+++ Sources/StreamChat/Workers/MessageUpdater.swift
@@ -1,5 +1,5 @@
 //
-// Copyright © 2023 Stream.io Inc. All rights reserved.
+// Copyright © 2024 Stream.io Inc. All rights reserved.
 //
 
 import CoreData
@@ -511,20 +511,13 @@ class MessageUpdater: Worker {
     /// - Parameters:
     ///   - id: The attachment identifier.
     ///   - completion: Called when the attachment database entity is updated. Called with `Error` if update fails.
-    func restartFailedAttachmentUploading(
+    func restartAttachmentUploading(
         with id: AttachmentId,
         completion: @escaping (Error?) -> Void
     ) {
         database.write({
             guard let attachmentDTO = $0.attachment(id: id) else {
                 throw ClientError.AttachmentDoesNotExist(id: id)
-            }
-
-            guard case .uploadingFailed = attachmentDTO.localState else {
-                throw ClientError.AttachmentEditing(
-                    id: id,
-                    reason: "uploading can be restarted for attachments in `.uploadingFailed` state only"
-                )
             }
 
             attachmentDTO.localState = .pendingUpload